% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addValue.R
\name{addvalue}
\alias{addvalue}
\title{Add values}
\usage{
addvalue(x, y)
}
\arguments{
\item{x}{one value}

\item{y}{another value}
}
\value{
numeric value
}
\description{
A function that add values
}
\details{

}
