% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{another_function}
\alias{another_function}
\title{This is my other function}
\usage{
another_function(param1)
}
\arguments{
\item{param1}{- The first parameter}
}
\value{
returns the number input as param 1
}
\description{
This is my other function
}
\examples{
another_function(10)
}
