% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_pair.R
\name{apa_t_pair}
\alias{apa_t_pair}
\title{APA Text for Paired-Samples T-Test}
\usage{
apa_t_pair(x, y, dv = "the DV", level1 = "level 1", level2 = "level 2")
}
\arguments{
\item{x}{A numeric vector of values for level 1.}

\item{y}{A numeric vector of values for level 2.}

\item{dv}{A character string describing the dependent variable (DV) in the output statement. Default is "the DV".}

\item{level1}{A character string describing level 1 in the output statement. Default is "level 1".}

\item{level2}{A character string describing level 2 in the output statement. Default is "level 2".}
}
\value{
A character string with the APA-formatted results of the paired-samples t-test.
}
\description{
Create APA-formatted text for the results of a paired-samples t-test.
}
\examples{
# Assuming self_res_att is a dataframe loaded in your environment
apa_t_pair(x = self_res_att$f_self, y = self_res_att$f_non)

# Specify the text for DV and levels
apa_t_pair(x = self_res_att$f_self, y = self_res_att$f_non,
           dv = "preferences for female faces",
           level1 = "participants who resembled those faces",
           level2 = "non-self participants")
}
