% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconversionCtoF.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{x}{The the Celsius temp that you would like to turn Fahrenheit}
}
\value{
The final Fahrenheit value.
}
\description{
This function will let you convert Celsius to Fahrenheit
}
