% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{check_condition}
\alias{check_condition}
\title{Check your condition. Ar you tired/nervous or both?}
\usage{
check_condition(tired, nervous)
}
\arguments{
\item{tired}{are you tired?}

\item{nervous}{are you nervous?}
}
\value{
advice
}
\description{
Check your condition. Ar you tired/nervous or both?
}
\examples{
check_condition(tired=TRUE,nervous=TRUE)
}
