\name{corCoefFiles}
\alias{corCoefFiles}
\title{corCoefFiles}
\usage{
  corCoefFiles(firstPath,secondPath)
}
\arguments{
	\item{firstPath}{Path to the first csv file. Required.}
	\item{secondPath}{Path to the second csv file. Required.}
}
\description{
  Counting correlation coeficient and covariance of two given columns.
}