% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dependent_function}
\alias{dependent_function}
\title{A filtering function for mtcars}
\usage{
dependent_function(x = mtcars, n_cyl = 4)
}
\arguments{
\item{x}{- Actually mtcars}

\item{n_cyl}{- number of cyls}
}
\value{
filtered mtcars filtered by cylinders
}
\description{
A filtering function for mtcars
}
\examples{
dependent_function(mtcars, 4)
}
