\name{drawQQ}
\alias{drawQQ}
\title{drawQQ}
\usage{
  drawQQ(vec, distribution)
}
\arguments{
	\item{vec}{Numeric vector containing data. Required.}
	\item{distribution}{Chosen distribution to compare with.}
}
\description{
   Function for drawing Q-Q plot for given data and given theoretical distribution to compare with.
}