\name{edfPlot}
\alias{edfPlot}
\title{edfPlot}
\usage{
  edfPlot(vec)
}
\arguments{
	\item{vec}{Numeric vector containing data. Required.}
}
\description{
  Function making graph of the empirical distribution function from given data. Function draws lines of theoretical distributions to the edf graph to compare with.
}