% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_function}
\alias{example_function}
\title{Example Function}
\usage{
example_function(param1, param2)
}
\arguments{
\item{Parameter}{1, does parameter like things}

\item{Parameter}{2, does other parameter like things}
}
\value{
Always gives back the number 3
}
\description{
Example Function
}
\examples{
example_function()
}
