% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getextent.R
\name{getextent}
\alias{getextent}
\title{getextent}
\usage{
getextent()
}
\value{
returns list with extent files in working directory which can be used as bbox
}
\description{
creates list of extents drawn in leaflet
}
\details{
uses leaflet to draw polygon
}
\note{
experimental, just checking how to map a package and how to use gitlab
}
\author{
Jouke <joukeverstappen@rhdhv.com>
}
