\name{KSTest}
\alias{KSTest}
\title{KSTest}
\usage{
  KSTest(vec, distribution, alpha)
}
\arguments{
	\item{vec}{Numeric vector containing data. Required.}
	\item{distribution}{Given distribution to compare with. Required.}
	\item{alpha}{Given sagnificance level fot the test. Required.}
}
\description{
   Function for One-sample Kolmogorov-Smirnov test. 
}