\name{KSTwoSam}
\alias{KSTwoSam}
\title{KSTwoSam}
\usage{
  KSTwoSam(firstPath,secondPath,alpha)
}
\arguments{
	\item{firstPath}{Path to the first csv file. Required.}
	\item{secondPath}{Path to the second csv file. Required.}
	\item{alpha}{Sagnificance level of the test. Required.}
}
\description{
  Function for two-sample Kolmogorov-Smirnov test. 
}