% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_function.R
\name{ma_function}
\alias{ma_function}
\title{Ma super fonction}
\usage{
ma_function(x, na_rm = FALSE)
}
\arguments{
\item{x}{C'est un vecteur numerique dont on veut calculer la moyenne}
}
\value{
Un vecteur de longueur 1 qui est la moyenne de x
}
\description{
Ceci est ma super function mais la je detaille
}
\examples{
ma_function(1:3)
\dontrun {
ma_function(1:3)
}
}
