% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.R
\name{minus}
\alias{minus}
\title{Return the minus of two numbers}
\usage{
minus(a, b)
}
\arguments{
\item{a}{Number one}

\item{b}{Number two}
}
\value{
minus of two numbers
}
\description{
Return the minus of two numbers
}
\examples{
library(newPackage)
minus(2 , 5)
}
