% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{my_func}
\alias{my_func}
\title{A function that can multiple and provide a text output}
\usage{
my_func(one, two, text_1)
}
\arguments{
\item{one}{first value  (numeric only)}

\item{two}{second value (numeric only)}

\item{text_1}{text (string only)}
}
\description{
A function that can multiple and provide a text output
}
\examples{
my_func(1,2,"my age is")
}
