% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myFuncName.R
\name{myFuncName}
\alias{myFuncName}
\title{A function quick description}
\usage{
myFuncName(param1, param2 = "Blah", ...)
}
\arguments{
\item{param1}{Info about param1 e.g. data type, guidance}

\item{param2}{Info about param1 e.g. data type, guidance}

\item{...}{Additional values to pass to x, y, z}
}
\value{
returnDT Info about what is returned by the function
}
\description{
A more detailed description that can span multiple lines
for readability. Covers concepts, typical usages etc.
}
\examples{
# Sample code that illustrates usage
}
\keyword{allowing}
\keyword{search}
\keyword{words}

