% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{package_data}
\alias{package_data}
\title{Runoff data}
\format{
a data frame with
\itemize{
\item fert_application
\item land_area
\item precipitation
\item water_quality
\item nutrient concentration
}
}
\usage{
package_data
}
\description{
A data fram of the inputs for runoff_rates function including #' fertilizer application rates, land area, and precipitation
}
\keyword{datasets}
