\name{pairedTTest}
\alias{pairedTTest}
\title{pairedTTest}
\usage{
  pairedTTest(firstPath, secondPath, alpha)
}
\arguments{
	\item{firstPath}{Path to the first csv file. Required.}
	\item{secondPath}{Path to the second csv file. Required.}
	\item{alpha}{Confidence level.}
}
\description{
   Function for two-tailed test of population mean. 
}