% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimmonsMons.R
\name{pokeDex}
\alias{pokeDex}
\title{Function for getting data about a single pokemon}
\usage{
pokeDex(name)
}
\arguments{
\item{name}{description}
}
\value{
None
}
\description{
Function for getting data about a single pokemon
}
\examples{
pokeDex("Litwick")
}
