% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimmonsMons.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{Data about Pokemon}
\format{
A datafram wit 1076 number of rows/observations and 47 varibles
\describe {
\item{Pokemon.Id}{the ID number for this pokemon in this dataset}
\item{Pokemon.Name}{The official name of the pokemon}
}
}
\usage{
pokemon
}
\description{
a dataset with data about pokemon from source
}
\examples{
pokemon
}
\keyword{datasets}
