% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_draw.R
\name{polygon_draw}
\alias{polygon_draw}
\title{polygon_draw}
\usage{
polygon_draw()
}
\value{
returns list with extent files in working directory which can be used as bbox
}
\description{
creates list of extents drawn in leaflet
}
\details{
uses leaflet to draw polygon
}
\note{
experimental, just checking how to map a package and how to use gitlab
}
\author{
Jouke <joukeverstappen@rhdhv.com>
}
