% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygondraw.R
\docType{package}
\name{polygondraw}
\alias{polygondraw}
\alias{polygondraw-package}
\title{getextent: get extent from drawn polygon}
\description{
Easily create \code{extent} files for creating bbox
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
