% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print_emoji}
\alias{print_emoji}
\alias{print_emoji.data.frame}
\title{The function prints emoji before dataframe}
\usage{
print_emoji(x, ...)

\method{print_emoji}{data.frame}(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{aurgument passes to \code{\link{print}}}
}
\description{
The function prints emoji before dataframe
}
\examples{
print_emoji(mtcars)
}
