% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Round a Number}
\usage{
round0(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector to be rounded.}

\item{digits}{An integer specifying the number of decimal places to round to. Default is 0.}
}
\value{
A character vector with the rounded numbers as strings.
}
\description{
Rounds a number to the specified number of decimal places and returns it as a string.
}
\examples{
round0(3.14159)
round0(3.14159, 2)
}
