% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_rates.R
\name{runoff_rates}
\alias{runoff_rates}
\title{Compute Nitrogen concentration in runoff in given data set}
\usage{
runoff_rates(land_area, application_rate, precipitation)
}
\arguments{
\item{land_area}{(ha) size of land where fertilizer is applied}

\item{application_rate}{kg Nitrogen/ha}

\item{runoff_coeff}{constant based on soil type}

\item{nutrient_conc}{concentration of nitrogen in land area (mg/L)}

\item{precipitaion}{rainfall intensity (mL/24hr)}

\item{nitrogen_max}{(mg/L) maximum nitrogen level in fresh water bodies by EPA (using NJ standards)}
}
\value{
runoff concentration
}
\description{
This function allows the user to input land area, fertlizer application rate, and precipitation and compute the estimated
amount of water into the nearby waterbody and how run off impacts water quality.
}
