% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/sisbid.R
\docType{package}
\name{sisbid}
\alias{sisbid}
\alias{sisbid-package}
\alias{sisbid}
\alias{sisbid-package}
\title{Basic Compilation of RR scripts}
\description{
This package serves to collect scripts and templates for producing reports in a more reproducible fashion.

This package serves to collect scripts and templates for producing reports in a more reproducible fashion.
}
