% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby.R
\name{sortby}
\alias{sortby}
\title{sort a data frame}
\usage{
sortby(df, col, desc = FALSE)
}
\arguments{
\item{df}{The data.frame to be sorted}

\item{col}{The column name (entered as a string) by which to sort the data.frame}

\item{desc}{Do you want it sorted in descending order? Defaults to FLASE.}
}
\value{
The sorted dataframe
}
\description{
This function accepts a data.frame and column number
  to sort the data.frame by that columns.
}
\examples{
subject <- c(1:10)
height <- round(rnorm(10, mean = 66, sd = 3),1)
d <- data.frame(subject, height)

d2 <- sortby(d, height)
d2

d3 <- sortby(d, height, desc = TRUE)
d3

}
