% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_head.R
\name{top}
\alias{top}
\title{Show the top part of an object}
\usage{
top(mat, n = 5)
}
\arguments{
\item{mat}{a matrix or data.frame}

\item{n}{number of rows/ columns to display}
}
\value{
A matrix with the specified number of rows/columns
}
\description{
Combines head with a number of columns arguments
}
\examples{
\dontrun{ x = matrix(rnorm(10),ncol = 5)
             top(x, n = 2)
}
}
\seealso{
\code{ \link{head}}
}
