\name{twoTailedTTest}
\alias{twoTailedTTest}
\title{twoTailedTTest}
\usage{
  twoTailedTTest(vec, mu0, alpha)
}
\arguments{
	\item{vec}{Numeric vector containing data. Required.}
	\item{mu0}{Hypothesis value of mu. Required.}
	\item{alpha}{Confidence level.}
}
\description{
   Function for two-tailed test of population mean. 
}