\name{drawHist}
\alias{drawHist}
\title{drawHist}
\usage{
  drawHist(vec, widthOfBin)
}
\arguments{
	\item{vec}{Numeric vector containing data. Required.}
	\item{widthOfBin}{Width of histogram bin given from user. Required.}
}
\description{
   Function for drawing histogram with given width of bin.
}