\name{edfPlot}
\alias{edfPlot}
\title{edfPlot}
\usage{
  edfPlot(vec, distribution)
}
\arguments{
	\item{vec}{Numeric vector containing data. Required.}
	\item{distribution}{Chosen distribution to compare with.}
}
\description{
  Function making graph of the empirical distribution function from given data. If distribution value is given, function draws line of this distribution to the edf graph to compare with.
}