% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpackage-package.R
\docType{package}
\name{newpackage-package}
\alias{newpackage}
\alias{newpackage-package}
\title{newpackage: My wonderful pacakage}
\description{
Nothing. It is just an exercice. Can include everything 
             <https://google.com>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MaelDore/newpackage}
  \item Report bugs at \url{https://github.com/MaelDore/newpackage/issues}
}

}
\author{
\strong{Maintainer}: Mael Dore \email{mael.dore@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
