\name{sumFileData}
\alias{sumFileData}
\title{sumFileData}
\usage{
  sumFileData(path)
}
\arguments{
	\item{path}{Path to the csv file. Required.}
	\item{var}{Name of the variable of given column. Required.}
}
\description{
  Third try to analyze some data from file...with dynamic path to file given by argument and trying to give the data to numeric vector.
}